refactor(vectordb): split lancedb.ts into focused modules
Splits monolithic 1119-line file into smaller, focused modules:
- lancedb.ts: Main VectorDB class (~200 lines)
- query.ts: Search operations (search, scanWithFilter, querySymbols) (~490 lines)
- batch-insert.ts: Batch insertion with retry logic (~160 lines)
- maintenance.ts: CRUD operations (clear, deleteByFile, updateFile) (~70 lines)

Benefits:
- All files now under 500 lines, parseable by Tree-sitter AST
- Better code navigation and organization
- Clearer separation of concerns
- Easier to maintain and test individual modules
